==> 2_SNP_analyses_b37.Rmd <==
---
title: "Genetic analyses"
subtitle: Accompanying 'MR_CVD_MDD'
author: '[Sander W. van der Laan, PhD](https://vanderlaanand.science) | s.w.vanderlaan@gmail.com'
date: '`r Sys.Date()`'
output:
  html_notebook: 
    cache: yes
    code_folding: hide
    collapse: yes
    df_print: paged
    fig.align: center
    fig_caption: yes
    fig_height: 10
    fig_retina: 2
    fig_width: 12
    theme: paper
    toc: yes
    toc_float:
      collapsed: no
      smooth_scroll: yes
    highlight: tango
mainfont: Helvetica
editor_options:
  chunk_output_type: inline
bibliography: references.bib
knit: worcs::cite_all
---

# General Setup

```{r echo = FALSE}
rm(list = ls())
```

```{r LocalSystem, echo = FALSE}
source("scripts/local.system.R")
```

```{r Source functions}
source("scripts/functions.R")
```

```{r loading_packages, message=FALSE, warning=FALSE}
source("scripts/pack01.packages.R")
```

```{r Setting: Colors}

Today = format(as.Date(as.POSIXlt(Sys.time())), "%Y%m%d")
Today.Report = format(as.Date(as.POSIXlt(Sys.time())), "%A, %B %d, %Y")

source("scripts/colors.R")
```

```{r setup_notebook, include=FALSE}
# We recommend that you prepare your raw data for analysis in 'prepare_data.R',
# and end that file with either open_data(yourdata), or closed_data(yourdata).
# Then, uncomment the line below to load the original or synthetic data
# (whichever is available), to allow anyone to reproduce your code:
# load_data()

# further define some knitr-options.
knitr::opts_chunk$set(fig.width = 12, fig.height = 8, fig.path = 'Figures/', 
                      warning = TRUE, # show warnings during codebook generation
                      message = TRUE, # show messages during codebook generation
                      error = TRUE, # do not interrupt codebook generation in case of errors, 
                                    # usually better for debugging
                      echo = TRUE,  # show R code
                      eval = TRUE)

ggplot2::theme_set(ggplot2::theme_minimal())
# pander::panderOptions("table.split.table", Inf)
library("worcs")
library("rmarkdown")
```


## Background

Collaboration to study `r TRAIT_OF_INTEREST` in relation to atherosclerotic plaques characteristics.

-   `Genes.xlsx` - list of genes of interest. 
-   `Variants.xlsx` - list of variant(s) of interest. 

```{r targets, message=FALSE, warning=FALSE}
library(openxlsx)

gene_list_df <- read.xlsx(paste0(TARGET_loc, "/targets.xlsx"), sheet = "Genes")

gene_list <- unlist(gene_list_df$Gene)
gene_list

variant_list <- read.xlsx(paste0(TARGET_loc, "/targets.xlsx"), sheet = "Variants")

DT::datatable(variant_list)

```


## This notebook 

In this notebook we setup the files for the SNP analysis. 

## Athero-Express Biobank Study

The [*Athero-Express Biobank Study (AE)*](http://www.atheroexpress.nl){target="_blank"} contains plaque material of
patients that underwent endarterectomyat two Dutch tertiary referral centers. Details of the study design were described
before. Briefly, blood and plaque material were obtained during endarterectomy and stored at -80 ℃. Only carotid
endarterectomy (CEA) patients were included in the present study. All patients provided informed consent and the study
was approved by the medical ethics committee.

## Athero-Express Genomics Study

### DNA isolation and genotyping

We genotyped the AE in three separate, but consecutive experiments. In short, DNA was extracted from EDTA blood or (when
no blood was available) plaque samples of 1,858 consecutive patients from the Athero-Express Biobank Study and genotyped
in 3 batches.

For the *Athero-Express Genomics Study 1 (AEGS1)* 891 patients (602 males, 262 females, 27 unknown sex), included
between 2002 and 2007, were genotyped (440,763 markers) using the Affymetrix Genome-Wide Human SNP Array 5.0 (SNP5) chip
(Affymetrix Inc., Santa Clara, CA, USA) at [Eurofins Genomics](https://www.eurofinsgenomics.eu/){target="_blank"}
(formerly known as AROS).

For the *Athero-Express Genomics Study 2 (AEGS2)* 954 patients (640 makes, 313 females, 1 unknown sex), included between
2002 and 2013, were genotyped (587,351 markers) using the Affymetrix AxiomⓇ GW CEU 1 Array (AxM) at the [Genome Analysis
Center](https://www.helmholtz-muenchen.de/no_cache/gac/index.html){target="_blank"}.

For the *Athero-Express Genomics Study 3 (AEGS3)* 658 patients (448 males, 203 females, 5 unknown sex), included between
2002 and 2016, were genotyped (693,931 markers) using the Illumina GSA MD v1 BeadArray (GSA) at [Human Genomics
Facility, HUGE-F](http://glimdna.org/index.html){target="_blank"}.

All experiments were carried out according to OECD standards.

### Genotyping calling

We used the genotyping calling algorithms as advised by Affymetrix (AEGS1 and AEGS2) and Illumina (AEGS3):

-   AEGS1: BRLMM-P
-   AEGS2: AxiomGT1
-   AEGS3: Illumina GenomeStudio

### Quality control after genotyping

After genotype calling, we adhered to community standard quality control and assurance (QCA) procedures of the genotype
data from AEGS1, AEGS2, and AEGS3. Samples with low average genotype calling and sex discrepancies (compared to the
clinical data available) were excluded. The data was further filtered on:

1)  individual (sample) call rate \> 97%,
2)  SNP call rate \> 97%,
3)  minor allele frequencies (MAF) \> 3%,
4)  average heterozygosity rate ± 3.0 s.d.,
5)  relatedness (pi-hat \> 0.20),
6)  Hardy--Weinberg Equilibrium (HWE p \< 1.0×10<sup>−3\<\sup\>), and
7)  Monomorphic SNPs (\< 1.0×10<sup>−6\<\sup\>).

After QCA 2,493 samples remained, 108 of non-European descent/ancestry, and 156 related pairs. These comprise 890
samples and 407,712 SNPs in AEGS1, 869 samples and 534,508 SNPs in AEGS2, and 649954 samples and 534,508 SNPs in AEGS3
remained.

### Imputation

Before phasing using SHAPEIT2, data was lifted to genome build b37 using the liftOver tool from UCSC
(<https://genome.ucsc.edu/cgi-bin/hgLiftOver>). Finally, data was imputed with 1000G phase 3, version 5 and HRC release
1.1 as a reference using the [Michigan Imputation Server](https://imputationserver.sph.umich.edu/){target="_blank"}.
These results were further integrated using QCTOOL v2, where HRC imputed variants are given precendence over 1000G phase
3 imputed variants.

### Quality control after imputation

We compared quality of the three AEGS datasets, and listed some variables of interest.

-   sample type (EDTA blood or plaque)
-   genotyping chip used
-   reason for filtering

We checked the studytype (AE or not), and *identity-by-descent (IBD)* within and between datasets to aid in sample
mixups, duplicate sample use, and relatedness. In addition, during genotyping quality control samples were identified
that deviated from *Hardy-Weinberg Equilibrium (HWE)*, had discordance in sex-coding and genotype sex, and deviated from
the *principal component analysis (PCA)* plot.

We will load the Athero-Express Biobank Study data, and all the samples that were send for genotyping and the final
QC'ed sampleList.

# Loading data

Loading Athero-Express Biobank Study clinical and biobank data, as well as the SampleList of genetic data. We simply load the previously saved `RDS`-file and extract the clinical data from that.

```{r LoadAEDB}
cat("* get Athero-Express Biobank Study Database...\n")

AEDB.CEA <- readRDS(file = paste0(OUT_loc, "/",Today,".",TRAIT_OF_INTEREST,".AEDB.CEA.RDS"))
# AEDB.CEA <- readRDS(file = paste0(OUT_loc, "/20240109.",TRAIT_OF_INTEREST,".AEDB.CEA.RDS"))
AEDB.CEA[1:10, 1:10]
dim(AEDB.CEA)

AEDB.full <- readRDS(file = paste0(OUT_loc, "/",Today,".",TRAIT_OF_INTEREST,".AEDB.FULL.RDS"))
# AEDB.full <- readRDS(file = paste0(OUT_loc, "/20240109.",TRAIT_OF_INTEREST,".AEDB.FULL.RDS"))
AEDB.full[1:10, 1:10]
dim(AEDB.full)


cat("\n* get Athero-Express Genomics Study keys...\n")
AEGS123.sampleList.keytable <- fread(paste0(AEGSQC_loc, "/QC/SELECTIONS/20200419.QC.AEGS123.sampleList.keytable.txt"))

dim(AEGS123.sampleList.keytable)
# AEGS123.sampleList.keytable[1:10, 1:10]
```


# Athero-Express Genomics Study

## Prepare baseline

Let's combine the full Athero-Express Biobank Study with the key-table containing the AEGS data.

> NOTE: this should sum to 2,124 samples with genotypes.

```{r create AEGS}
AEGS <- merge(AEDB.full, AEGS123.sampleList.keytable, by.x = "STUDY_NUMBER", by.y = "STUDY_NUMBER", sort = FALSE,
                  all = TRUE)

dim(AEGS)

AEGS$UPID.y <- NULL
names(AEGS)[names(AEGS) == "UPID.x"] <- "UPID"
AEGS$Age.y <- NULL
names(AEGS)[names(AEGS) == "Age.x"] <- "Age"

table(AEGS$CHIP, useNA = "ifany")

AEGS$GWAS <- AEGS$CHIP
AEGS$GWAS[is.na(AEGS$GWAS)] <- "not genotyped"
AEGS$GWAS[AEGS$GWAS != "not genotyped"] <- "genotyped"

table(AEGS$CHIP, AEGS$GWAS, useNA = "ifany")
```

Also a visualisation of the AEGS with AEDB overlaps.

```{r visualise AEGS overlaps, message=FALSE, warning=FALSE}
library(UpSetR)
require(ggplot2)
require(plyr)
require(gridExtra)
require(grid)

AEDB.availGWAS = list(
AEGS1 = subset(AEGS, CHIP == "AffySNP5", select = c("STUDY_NUMBER"))[,1],
AEGS2 = subset(AEGS, CHIP == "AffyAxiomCEU", select = c("STUDY_NUMBER"))[,1],
AEGS3 = subset(AEGS, CHIP == "IllGSA", select = c("STUDY_NUMBER"))[,1],
AEDB = AEGS$STUDY_NUMBER)

p1 <- UpSetR::upset(fromList(AEDB.availGWAS), 
                    sets = c("AEDB", "AEGS1", "AEGS2", "AEGS3"), 
                    main.bar.color = c(uithof_color[15], uithof_color[3], uithof_color[2], uithof_color[21]), 
                    mainbar.y.label	= "intersection sample size", 
                    sets.bar.color = c(uithof_color[15], uithof_color[2], uithof_color[3], uithof_color[21]), 
                    sets.x.label = "sample size", keep.order = TRUE)

p1

pdf(paste0(PLOT_loc, "/", Today, ".overlap.AEDB_AEGS123.UpSetR.pdf"))
  p1
dev.off()

png(paste0(PLOT_loc, "/", Today, ".overlap.AEDB_AEGS123.UpSetR.png"))
  p1
dev.off()

rm(p1)

```
> Please refer to the `AEDB.CEA.baseline.html` document for the details on the informed consent selection.

```{r SpecificSelection}
table(AEGS$Artery_summary, AEGS$QC2018_FILTER)
table(AEGS$informedconsent, AEGS$QC2018_FILTER)
AEGSselect <- subset(AEGS, 
                     informedconsent != "missing" & # we are really strict in selecting based on 'informed consent'!
                     informedconsent != "no, died" & 
                     informedconsent != "yes, no tissue, no commerical business" & 
                     informedconsent != "yes, no tissue, no questionnaires, no medical info, no commercial business" & 
                     informedconsent != "yes, no tissue, no questionnaires, no health treatment, no commerical business" & 
                     informedconsent != "yes, no tissue, no questionnaires, no health treatment, no medical info, no commercial business" & 
                     informedconsent != "yes, no tissue, no health treatment" & 
                     informedconsent != "yes, no tissue, no questionnaires" & 
                     informedconsent != "yes, no tissue, health treatment when possible" & 
                     informedconsent != "yes, no tissue" & 
                     informedconsent != "yes, no tissue, no questionnaires, no health treatment, no medical info" & 
                     informedconsent != "yes, no tissue, no questionnaires, no health treatment, no commercial business" & 
                     informedconsent != "no, doesn't want to" & 
                     informedconsent != "no, unable to sign" & 
                     informedconsent != "no, no reaction" & 
                     informedconsent != "no, lost" & 
                     informedconsent != "no, too old" & 
                     informedconsent != "no (never asked for IC because there was no tissue)" & 
                     informedconsent != "no, endpoint" & 
                     informedconsent != "wil niets invullen, wel alles gebruiken" & 
                     informedconsent != "nooit geincludeerd" & 
                     informedconsent != "yes, no DNA")

AEGSselect.CEA <- subset(AEGS, !is.na(QC2018_FILTER) & QC2018_FILTER != "issue" & QC2018_FILTER != "family_discard" &
                     (Artery_summary == "carotid (left & right)" | Artery_summary == "other carotid arteries (common, external)") & # we only want carotids
                     informedconsent != "missing" & # we are really strict in selecting based on 'informed consent'!
                     informedconsent != "no, died" & 
                     informedconsent != "yes, no tissue, no commerical business" & 
                     informedconsent != "yes, no tissue, no questionnaires, no medical info, no commercial business" & 
                     informedconsent != "yes, no tissue, no questionnaires, no health treatment, no commerical business" & 
                     informedconsent != "yes, no tissue, no questionnaires, no health treatment, no medical info, no commercial business" & 
                     informedconsent != "yes, no tissue, no health treatment" & 
                     informedconsent != "yes, no tissue, no questionnaires" & 
                     informedconsent != "yes, no tissue, health treatment when possible" & 
                     informedconsent != "yes, no tissue" & 
                     informedconsent != "yes, no tissue, no questionnaires, no health treatment, no medical info" & 
                     informedconsent != "yes, no tissue, no questionnaires, no health treatment, no commercial business" & 
                     informedconsent != "no, doesn't want to" & 
                     informedconsent != "no, unable to sign" & 
                     informedconsent != "no, no reaction" & 
                     informedconsent != "no, lost" & 
                     informedconsent != "no, too old" & 
                     informedconsent != "no (never asked for IC because there was no tissue)" & 
                     informedconsent != "no, endpoint" & 
                     informedconsent != "wil niets invullen, wel alles gebruiken" & 
                     informedconsent != "nooit geincludeerd" & 
                     informedconsent != "yes, no DNA")

dim(AEGSselect)

table(AEGSselect$Artery_summary, AEGSselect$QC2018_FILTER)
table(AEGSselect$Artery_summary, AEGSselect$CHIP)
table(AEGSselect$QC2018_FILTER, AEGSselect$CHIP)
table(AEGSselect$QC2018_FILTER, AEGSselect$SAMPLE_TYPE)

AEDB.temp <- subset(AEGSselect,  select = c("STUDY_NUMBER", "UPID", "Age", "Gender", "Hospital", "Artery_summary", "QC2018_FILTER", "CHIP", "SAMPLE_TYPE"))
require(labelled)
AEDB.temp$Gender <- to_factor(AEDB.temp$Gender)
AEDB.temp$Hospital <- to_factor(AEDB.temp$Hospital)
AEDB.temp$Artery_summary <- to_factor(AEDB.temp$Artery_summary)
AEDB.temp$QC2018_FILTER <- to_factor(AEDB.temp$QC2018_FILTER)
AEDB.temp$CHIP <- to_factor(AEDB.temp$CHIP)
AEDB.temp$SAMPLE_TYPE <- to_factor(AEDB.temp$SAMPLE_TYPE)

DT::datatable(AEDB.temp[1:10,], caption = "Excerpt of the whole AEDB.", rownames = FALSE)

rm(AEDB.temp)


```

## Athero-Express Genomics Study Baseline Characteristics

Showing the baseline table of the Athero-Express Genomics Study.

```{r Baseline SampleSelect: prepare}
# Create baseline tables
# http://rstudio-pubs-static.s3.amazonaws.com/13321_da314633db924dc78986a850813a50d5.html
AEGSselect$GWAS <- to_factor(AEGSselect$GWAS)
AEGSselect$CHIP <- to_factor(AEGSselect$CHIP)
AEGSselect$PCA <- to_factor(AEGSselect$PCA)
AEGSselect$SAMPLE_TYPE <- to_factor(AEGSselect$SAMPLE_TYPE)
AEGSselect$informedconsent <- to_factor(AEGSselect$informedconsent)
AEGSselect$Artery_summary <- to_factor(AEGSselect$Artery_summary)

AEGSselect.CEA$GWAS <- to_factor(AEGSselect.CEA$GWAS)
AEGSselect.CEA$CHIP <- to_factor(AEGSselect.CEA$CHIP)
AEGSselect.CEA$PCA <- to_factor(AEGSselect.CEA$PCA)
AEGSselect.CEA$SAMPLE_TYPE <- to_factor(AEGSselect.CEA$SAMPLE_TYPE)
AEGSselect.CEA$informedconsent <- to_factor(AEGSselect.CEA$informedconsent)
AEGSselect.CEA$Artery_summary <- to_factor(AEGSselect.CEA$Artery_summary)

cat("===========================================================================================\n")
cat("CREATE BASELINE TABLE\n")

# Baseline table variables
basetable_vars = c("Hospital", 
                   "Age", "Gender", 
                   "TC_final", "LDL_final", "HDL_final", "TG_final", 
                   "systolic", "diastoli", "GFR_MDRD", "BMI", 
                   "KDOQI", "BMI_WHO", 
                   "SmokerCurrent", "eCigarettes", "ePackYearsSmoking",
                   "DiabetesStatus", "Hypertension.selfreport", "Hypertension.selfreportdrug", "Hypertension.composite", 
                   "Hypertension.drugs", "Med.anticoagulants", "Med.all.antiplatelet", "Med.Statin.LLD", 
                   "Stroke_Dx", "sympt", "Symptoms.5G", "restenos",
                   "EP_composite", "EP_composite_time",
                   "macmean0", "smcmean0", "Macrophages.bin", "SMC.bin", "neutrophils", "Mast_cells_plaque", "vessel_density_averaged",
                   "IPH.bin", 
                   "Calc.bin", "Collagen.bin", 
                   "Fat.bin_10", "Fat.bin_40", "OverallPlaquePhenotype", "Plaque_Vulnerability_Index",
                   "SMC_rankNorm", "MAC_rankNorm", "Neutrophils_rankNorm", "MastCells_rankNorm", "VesselDensity_rankNorm",
                   "GWAS", "CHIP", "PCA", "Artery_summary",
                   "PCSK9_plasma", "PCSK9_plasma_rankNorm")

basetable_bin = c("Gender", 
                  "KDOQI", "BMI_WHO", 
                  "SmokerCurrent", 
                  "DiabetesStatus", "Hypertension.selfreport", "Hypertension.selfreportdrug", "Hypertension.composite", 
                  "Hypertension.drugs", "Med.anticoagulants", "Med.all.antiplatelet", "Med.Statin.LLD", 
                  "Stroke_Dx", "sympt", "Symptoms.5G", "restenos",
                  "EP_composite", "Macrophages.bin", "SMC.bin",
                  "IPH.bin", 
                  "Calc.bin", "Collagen.bin", 
                  "Fat.bin_10", "Fat.bin_40", "OverallPlaquePhenotype", "Plaque_Vulnerability_Index",
                  "GWAS", "CHIP", "PCA", "Artery_summary")

basetable_bin

basetable_con = basetable_vars[!basetable_vars %in% basetable_bin]
basetable_con

```

All Athero-Express Genomics Study data (n = 2,011), compared to the *remaining*, \_un_genotyped Athero-Express Biobank
Study.

```{r Baseline SampleSelect: Visualize, whole}
cat("\n===========================================================================================\n")
cat("DISPLAY BASELINE TABLE\n")

AEGSselect.tableOne = print(CreateTableOne(vars = basetable_vars, 
                                         # factorVars = basetable_bin,
                                         strata = "GWAS",
                                         data = AEGSselect, includeNA = TRUE), 
                          nonnormal = c(), missing = TRUE,
                          quote = FALSE, noSpaces = FALSE, showAllLevels = TRUE, explain = TRUE, 
                          format = "pf", 
                          contDigits = 3)[,1:6]
```


Baseline of the valid, CEA and genotyped data.

```{r Baseline SampleSelect: Visualize, CEA}
AEGSselect.CEA.tableOne = print(CreateTableOne(vars = basetable_vars, 
                                         # factorVars = basetable_bin,
                                         strata = "Gender",
                                         data = AEGSselect.CEA, includeNA = TRUE), 
                          nonnormal = c(), missing = TRUE,
                          quote = FALSE, noSpaces = FALSE, showAllLevels = TRUE, explain = TRUE, 
                          format = "pf", 
                          contDigits = 3)[,1:6]
```

### Baseline writing

Let's save the baseline characteristics of the Athero-Express Genomics Study.

```{r Baseline SampleSelection: write AEGS}
# Write basetable
require(openxlsx)

write.xlsx(file = paste0(BASELINE_loc, "/",Today,".",PROJECTNAME,".AEGS.BaselineTable.xlsx"), 
           format(as.data.frame(AEGSselect.tableOne), digits = 5, scientific = FALSE),
           rowNames = TRUE, 
           colNames = TRUE, 
           sheetName = "AEGS_Base_AEDB", overwrite = TRUE)

write.xlsx(file = paste0(BASELINE_loc, "/",Today,".",PROJECTNAME,".AEGS.CEA.BaselineTable.xlsx"), 
           format(as.data.frame(AEGSselect.CEA.tableOne), digits = 5, scientific = FALSE),
           rowNames = TRUE, 
           colNames = TRUE, 
           sheetName = "AEGS_Base_CEA_sex", overwrite = TRUE)

```

# AE eQTL Study (b37, n=626) baseline characteristics

Here we will load the list of samples that have overlapping genotype and transcriptomic data, and we will create a baseline table for these samples. These samples are derived from the molQTL analyses done using [QTLToolKit]().

```{r}
library(openxlsx)

# update list
SampleList_eQTL_b37 <- read.xlsx(paste0(TARGET_loc, "/SampleList_eQTL_b37.xlsx"), sheet = "SampleList_eQTL_b37")

DT::datatable(SampleList_eQTL_b37)
```

Let's match this list with the AEGS dataset, as this will have the genotype SampleID. 

```{r}
# Matching the SampleList_eQTL_b37 with the AEGS dataset
# ID_1 = SampleID

# AEGS_eQTL <- merge(AEGSselect, SampleList_eQTL_b37, 
#                    by.x = "ID_1", 
#                    by.y = "SampleID", 
#                    sort = FALSE,
#                    all.y = TRUE)

AEGS_eQTL <- merge(AEGS, SampleList_eQTL_b37, 
                   by.x = "ID_1", 
                   by.y = "SampleID", 
                   sort = FALSE,
                   all.y = TRUE)
```

```{r Baseline SampleSelection: write Prepare}
library(gtsummary)
library(dplyr)

# Drop variables that have <2 levels across the full dataset 
# This is to prevent the message:
# The following errors were returned during `add_p()`:✖ For variable `GWAS` (`Gender`) and "statistic", "p.value", and "parameter" statistics: 'x' and 'y' must have at least 2 levels
clean_vars <- basetable_vars[sapply(AEGS_eQTL[basetable_vars], function(x) nlevels(factor(x)) > 1)]

# Create the summary table, stratified by Gender
AEGS_eQTL.sex.tbl <- AEGS_eQTL %>%
  select(all_of(clean_vars), Gender) %>%
  tbl_summary(
    by = Gender,
    missing = "ifany",
    statistic = all_continuous() ~ "{mean} ({sd})"
  ) %>%
  add_p(test.args = all_categorical() ~ list(simulate.p.value = TRUE)) %>%
  bold_labels()

# Create the summary table
AEGS_eQTL.tbl <- AEGS_eQTL %>%
  select(all_of(clean_vars)) %>%
  tbl_summary(
    # by = Gender,
    missing = "ifany",
    statistic = all_continuous() ~ "{mean} ({sd})"
  ) %>%
  add_n %>%
  bold_labels()

```

```{r Baseline SampleSelection: write Excel}
library(writexl)

# Convert to a clean data frame
AEGS_eQTL.sex.tbl_df <- AEGS_eQTL.sex.tbl %>% as_tibble()

# Save to Excel
write.xlsx(AEGS_eQTL.sex.tbl_df,
           file = paste0(BASELINE_loc, "/",Today,".",PROJECTNAME,".AEGS_eQTL.Gender.BaselineTable.xlsx"), 
           rowNames = FALSE, 
           colNames = TRUE, 
           sheetName = "AEGS_eQTL_gender", overwrite = TRUE)

# Convert to a clean data frame
AEGS_eQTL.tbl_df <- AEGS_eQTL.tbl %>% as_tibble()

# Save to Excel
write.xlsx(AEGS_eQTL.tbl_df,
           file = paste0(BASELINE_loc, "/",Today,".",PROJECTNAME,".AEGS_eQTL.BaselineTable.xlsx"), 
           rowNames = FALSE, 
           colNames = TRUE, 
           sheetName = "AEGS_eQTL", overwrite = TRUE)

```


```{r Baseline SampleSelection: write Word}
library(officer)
library(flextable)

# Format for Word output with Calibri font and size 9, stratified by Gender
ft <- gtsummary::as_flex_table(AEGS_eQTL.sex.tbl) %>%
  theme_vanilla() %>%
  autofit() %>%
  set_table_properties(layout = "autofit") %>%
  flextable::font(fontname = "Calibri", part = "all") %>%
  flextable::fontsize(size = 9, part = "all")

# Save to Word
doc <- read_docx() %>%
  body_add_par("Table 1. Baseline characteristics Athero-Express Biobank Study - eQTL data only (b37, n=626), stratified by Gender") %>%
  body_add_flextable(ft)
print(doc, target = paste0(BASELINE_loc, "/",Today,".",PROJECTNAME,".AEGS_eQTL.Gender.BaselineTable.docx"))

# Print message
message("Table exported successfully to *.AEGS_eQTL.Gender.BaselineTable.xlsx and *.AEGS_eQTL.Gender.BaselineTable.docx with Calibri font and size 9.")
rm(ft, doc)

# Format for Word output with Calibri font and size 9
ft <- gtsummary::as_flex_table(AEGS_eQTL.tbl) %>%
  theme_vanilla() %>%
  autofit() %>%
  set_table_properties(layout = "autofit") %>%
  flextable::font(fontname = "Calibri", part = "all") %>%
  flextable::fontsize(size = 9, part = "all")

# Save to Word
doc <- read_docx() %>%
  body_add_par("Table 1. Baseline characteristics Athero-Express Biobank Study - eQTL data only (b37, n=626)") %>%
  body_add_flextable(ft)
print(doc, target = paste0(BASELINE_loc, "/",Today,".",PROJECTNAME,".AEGS_eQTL.BaselineTable.docx"))

# Print message
message("Table exported successfully to *.AEGS_eQTL.BaselineTable.xlsx and *.AEGS_eQTL.BaselineTable.docx with Calibri font and size 9.")
rm(ft, doc)

```


Let's also display the baseline data.

```{r}
AEGS_eQTL.sex.tbl
```

```{r}
AEGS_eQTL.tbl
```


## Save AEGS_eQTL

Save the `AEGS_eQTL` dataset for further use downstream.

```{r}
# Save the AEGS_eQTL dataset
saveRDS(AEGS_eQTL, file = paste0(OUT_loc, "/",Today,".",PROJECTNAME,".AEGS_eQTL.RDS"))
```



# Session information

------------------------------------------------------------------------------------------------------------------------

    Version:      v1.3.1
    Last update:  2025-04-10
    Written by:   Sander W. van der Laan (s.w.vanderlaan-2[at]umcutrecht.nl).
    Description:  Script to get some Athero-Express Biobank Study baseline characteristics.
    Minimum requirements: R version 3.4.3 (2017-06-30) -- 'Single Candle', Mac OS X El Capitan

    Changes log
    * v1.3.1 Fixed an issue where the Word-document output was not working properly.
    * v1.3.0 Added baseline table writing of eQTL data subset. 
    * v1.2.2 Textual fixes.
    * v1.2.1 Fixed issue with baseline writing.
    * v1.2.0 Update to the study database. 
    * v1.1.0 Major update to WORCS system. 
    * v1.0.6 Small bug fixes.
    * v1.0.5 Added png for overlap-figure.
    * v1.0.5 Removed obsolete references to objects.
    * v1.0.4 Fixed a mistake in the chr X sample-file creation. Now the order matches the chr X data.
    * v1.0.3 Fixed weight of files (limit of 10Mb per file for templates). Renamed entire repo.
    * v1.0.2 Added sex-specific .sample-files. Added GWASToolKit input-files.
    * v1.0.0 Initial version. Add 'plaque vulnerability index', Fixed baseline table, added codes, and results. Created sample-files.

------------------------------------------------------------------------------------------------------------------------

```{r eval = TRUE}
sessionInfo()
```

# Saving environment

```{r Saving}

save.image(paste0(PROJECT_loc, "/",Today,".",PROJECTNAME,".SNP_analyses.RData"))

```

+-----------------------------------------------------------------------------------------------------------------------------------------+
| <sup>© 1979-2025 Sander W. van der Laan | s.w.vanderlaan[at]gmail.com | [vanderlaanand.science](https://vanderlaanand.science).</sup> |
+-----------------------------------------------------------------------------------------------------------------------------------------+
